requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
python-decouple
psycopg2-binary
google-auth
requests
python-chess

# Real-time and WebSocket
channels>=4.0,<5.0
daphne>=4.0,<5.0
channels-redis>=4.0,<5.0

# Async
asgiref>=3.7,<4.0
stripe
python-dotenv
